major restructuring in the project: added core library
