+++ .gitignore
@@ -3,6 +3,7 @@ auto-report_stack-trace.txt
 
 # For beta purposes
 cogs/fun.py
+cogs/images.py
 cogs/globalban.py
 cogs/globalban.i18n.json
 cogs/handlers.py
