+++ cogs/basic.py
@@ -1,20 +1,16 @@
-from __future__ import annotations
-
 from time import perf_counter
-from typing import TYPE_CHECKING
 
 from discord.ext import commands
 
-if TYPE_CHECKING:
-	from main import Context, MyClient
+from core import Context, MyClient
 
 
 class Basic(commands.Cog, name="Basic"):
-	def __init__(self, client: "MyClient"):
+	def __init__(self, client: MyClient):
 		self.client = client
 
 	@commands.hybrid_command(name="ping", description="ping_specs-description")
-	async def ping(self, ctx: "Context"):
+	async def ping(self, ctx: Context):
 		# Database ping calculation
 		database_start = perf_counter()
 		await self.client.db.execute("SELECT 1")
@@ -23,5 +19,5 @@ class Basic(commands.Cog, name="Basic"):
 		await ctx.send("ping", latency=float(self.client.latency), db=float(database))
 
 
-async def setup(client: "MyClient"):
+async def setup(client: MyClient):
 	await client.add_cog(Basic(client))
