+++ cogs/mod.py
@@ -1,15 +1,14 @@
-from __future__ import annotations
-
 import datetime
 from copy import deepcopy
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Literal, Self
+from typing import Any, Literal, Self
 
 import asyncpg
 import discord
 from discord import app_commands
 from discord.ext import commands
 
+from core import Context, MyClient
 from helpers import (
 	CustomGuild,
 	CustomMember,
@@ -21,10 +20,6 @@ from helpers import (
 	seconds_to_text,
 	text_to_seconds,
 )
-from main import MyClient
-
-if TYPE_CHECKING:
-	from main import Context
 
 
 class CaseType(Enum):
@@ -627,7 +622,7 @@ class Ban(Case):
 @commands.guild_only()
 @app_commands.guild_only()
 class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
-	def __init__(self, client: "MyClient") -> None:
+	def __init__(self, client: MyClient) -> None:
 		self.client = client
 		self.custom_response = custom_response.CustomResponse(client, "mod")
 
@@ -676,7 +671,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	@commands.has_permissions(moderate_members=True)
 	async def warn(
 		self,
-		ctx: "Context",
+		ctx: Context,
 		user: discord.Member,
 		expires: str = None,
 		*,
@@ -737,7 +732,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	@commands.has_permissions(moderate_members=True)
 	async def mute(
 		self,
-		ctx: "Context",
+		ctx: Context,
 		user: discord.Member,
 		expires: str,
 		*,
@@ -774,7 +769,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	@app_commands.describe(user="unmute_specs-args-user-description")
 	@app_commands.checks.has_permissions(moderate_members=True)
 	@commands.has_permissions(moderate_members=True)
-	async def unmute(self, ctx: "Context", user: discord.Member):
+	async def unmute(self, ctx: Context, user: discord.Member):
 		if user.timed_out_until:
 			cases = await Mute.from_db(
 				self.client.db,
@@ -802,7 +797,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	)
 	@app_commands.checks.has_permissions(kick_members=True)
 	@commands.has_permissions(kick_members=True)
-	async def kick(self, ctx: "Context", user: discord.Member, *, reason: str = None):
+	async def kick(self, ctx: Context, user: discord.Member, *, reason: str = None):
 		if user == ctx.me:
 			await ctx.send("mod.kick.errors.bot")
 			return
@@ -835,7 +830,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	@commands.has_permissions(ban_members=True)
 	async def ban(
 		self,
-		ctx: "Context",
+		ctx: Context,
 		user: discord.User,
 		expires: str = None,
 		*,
@@ -870,7 +865,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	@app_commands.describe(user="unban_specs-args-user-description")
 	@app_commands.checks.has_permissions(ban_members=True)
 	@commands.has_permissions(ban_members=True)
-	async def unban(self, ctx: "Context", user: discord.User):
+	async def unban(self, ctx: Context, user: discord.User):
 		cases = await Ban.from_db(self.client.db, self.client, ctx.guild, user=user)
 		if cases:
 			for case in cases:
@@ -891,7 +886,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 	@commands.hybrid_command(name="slowmode", description="sm_specs-description", usage="sm_specs-usage")
 	@app_commands.describe(duration="sm_specs-args-duration-description", channel="sm_specs-args-channel-description")
 	@app_commands.rename(duration="sm_specs-args-duration-name", channel="sm_specs-args-channel-name")
-	async def slowmode(self, ctx: "Context", duration: str = None, channel: discord.TextChannel = None):
+	async def slowmode(self, ctx: Context, duration: str = None, channel: discord.TextChannel = None):
 		if not duration:
 			await ctx.send("mod.slowmode.current_slowmode", channel=CustomTextChannel.from_channel(ctx.channel))
 			return
@@ -920,7 +915,7 @@ class Moderation(commands.GroupCog, name="Moderation", group_name="mod"):
 @commands.guild_only()
 @app_commands.guild_only()
 class Cases(commands.Cog, name="Cases"):
-	def __init__(self, client: "MyClient") -> None:
+	def __init__(self, client: MyClient) -> None:
 		self.client = client
 		self.custom_response = custom_response.CustomResponse(client, "mod")
 
@@ -932,7 +927,7 @@ class Cases(commands.Cog, name="Cases"):
 	)
 	@app_commands.rename(case_id="caseinfo_specs-args-case_id-name")
 	@app_commands.describe(case_id="caseinfo_specs-args-case_id-description")
-	async def case(self, ctx: "Context", case_id: str):
+	async def case(self, ctx: Context, case_id: str):
 		try:
 			case_id = int(case_id)
 		except ValueError:
@@ -962,7 +957,7 @@ class Cases(commands.Cog, name="Cases"):
 	@app_commands.rename(case_id="casedel_specs-args-case_id-name")
 	@app_commands.checks.has_permissions(moderate_members=True)
 	@commands.has_permissions(moderate_members=True)
-	async def delete(self, ctx: "Context", case_id: str):
+	async def delete(self, ctx: Context, case_id: str):
 		try:
 			# because discord's app commands only support int up to 2^54, but discord snowflakes are 2^64,
 			# we need to convert the case id to an int ourselves :(
@@ -1016,7 +1011,7 @@ class Cases(commands.Cog, name="Cases"):
 	@commands.has_permissions(moderate_members=True)
 	async def edit(
 		self,
-		ctx: "Context",
+		ctx: Context,
 		case_id: str,
 		value: Literal["expires", "reason", "message"],
 		*,
@@ -1051,7 +1046,7 @@ class Cases(commands.Cog, name="Cases"):
 	)
 	@app_commands.describe(user="caselist_specs-args-user-description")
 	@app_commands.rename(user="caselist_specs-args-user-name")
-	async def list(self, ctx: "Context", user: discord.Member = None):
+	async def list(self, ctx: Context, user: discord.Member = None):
 		user = user or ctx.author
 
 		cases = await Case.from_user(self.client.db, user, self.client, ctx.guild, 10)
@@ -1088,6 +1083,6 @@ class Cases(commands.Cog, name="Cases"):
 		await ctx.send(**message)
 
 
-async def setup(client: "MyClient"):
+async def setup(client: MyClient):
 	await client.add_cog(Moderation(client))
 	await client.add_cog(Cases(client))
