+++ pyproject.toml
@@ -23,6 +23,9 @@ dependencies = [
     "pypokedex>=1.6.0",
     "art>=6.5",
     "ruff>=0.13.1",
+    "pillow>=11.3.0",
+    "asyncpraw>=7.8.1",
+    "pytemperature>=1.1",
 ]
 
 [tool.setuptools]
